+++ lib/common/spi_functions.h
@@ -12,10 +12,10 @@ extern uint16_t Vu;
 
 void     set_txdata_with(uint8_t tx_data[4], const adc_param data);
 void     split_uint32_to_uint8_array(const uint32_t data, uint16_t receiver[2]);
-uint16_t get_mmode_value(uint8_t Lgain, uint8_t Ngain, uint8_t LNsel,
-                         uint8_t DisHPF, uint8_t Amod, uint8_t Rmod,
-                         uint8_t Zxcon, uint8_t Pthresh);
-uint16_t get_average_value( const uint16_t data_array[], const uint8_t array_size);
+uint16_t get_mmode_value(uint16_t Lgain, uint16_t Ngain, uint16_t LNsel,
+                         uint16_t DisHPF, uint16_t Amod, uint16_t Rmod,
+                         uint16_t Zxcon, uint16_t Pthresh);
+uint16_t get_average_value(const uint16_t data_array[], const uint8_t array_size);
 
 uint16_t get_threshold(const uint8_t K); //K in per mille
